refactor default updater, subclass for hopkins and harvard
