+++ pass-grant-data/src/test/java/org/dataconservancy/pass/grant/data/JhuPassUpdaterTest.java
@@ -52,7 +52,7 @@ public class JhuPassUpdaterTest {
 
     private PassEntityUtil passEntityUtil = new CoeusPassEntityUtil();
 
-    private String domain = "johnshopkins.edu";
+   // private String domain = "johnshopkins.edu";
 
     private URI grantUri;
 
@@ -167,7 +167,7 @@ public class JhuPassUpdaterTest {
 
         resultSet.add(rowMap);
 
-        JhuPassUpdater passUpdater = new JhuPassUpdater(passClientMock, passEntityUtil, domain);
+        JhuPassUpdater passUpdater = new JhuPassUpdater(passClientMock);
         passUpdater.updatePass(resultSet, "grant");
 
         Map<URI, Grant> grantMap = passUpdater.getGrantUriMap();
@@ -204,7 +204,7 @@ public class JhuPassUpdaterTest {
         rowMap.put(C_USER_HOPKINS_ID, "A1A1A1");
         rowMap.put(C_UPDATE_TIMESTAMP, "2018-01-01 0:00:00.0");
 
-        JhuPassUpdater passUpdater = new JhuPassUpdater(passClientMock, passEntityUtil, domain);
+        JhuPassUpdater passUpdater = new JhuPassUpdater(passClientMock);
         User newUser = passUpdater.buildUser(rowMap);
 
         //unusual fields
@@ -223,7 +223,7 @@ public class JhuPassUpdaterTest {
         rowMap.put(C_PRIMARY_FUNDER_LOCAL_KEY, "8675309");
         rowMap.put(C_PRIMARY_FUNDER_POLICY, "policy1");
 
-        JhuPassUpdater passUpdater = new JhuPassUpdater(passClientMock, passEntityUtil, domain);
+        JhuPassUpdater passUpdater = new JhuPassUpdater(passClientMock);
         Funder newFunder = passUpdater.buildPrimaryFunder(rowMap);
 
         assertEquals("Funder Name", newFunder.getName());
@@ -240,7 +240,7 @@ public class JhuPassUpdaterTest {
         grantResultSet.add(rowMap);
 
         PassClient passClient = PassClientFactory.getPassClient();
-        JhuPassUpdater passUpdater = new JhuPassUpdater(passClient, passEntityUtil, domain);
+        JhuPassUpdater passUpdater = new JhuPassUpdater(passClient);
 
         passUpdater.updatePass(grantResultSet, "user");
 
@@ -254,7 +254,7 @@ public class JhuPassUpdaterTest {
         userResultSet.add(rowMap);
 
         PassClient passClient = PassClientFactory.getPassClient();
-        JhuPassUpdater passUpdater = new JhuPassUpdater(passClient, passEntityUtil, domain );
+        JhuPassUpdater passUpdater = new JhuPassUpdater(passClient);
 
         passUpdater.updatePass(userResultSet, "grant");
